+++ src/Manifolds.jl
@@ -444,6 +444,7 @@ include("manifolds/Lorentz.jl")
 include("manifolds/CenteredMatrices.jl")
 include("manifolds/CholeskySpace.jl")
 include("manifolds/Circle.jl")
+include("manifolds/DeterminantOneMatrices.jl")
 include("manifolds/Elliptope.jl")
 include("manifolds/EmbeddedTorus.jl")
 include("manifolds/FixedRankMatrices.jl")
@@ -642,6 +643,7 @@ export Euclidean,
     CenteredMatrices,
     CholeskySpace,
     Circle,
+    DeterminantOneMatrices,
     Elliptope,
     EssentialManifold,
     FixedRankMatrices,