+++ src/groups/orthogonal.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-    Orthogonal{T} = GeneralUnitaryMultiplicationGroup{T,ℝ,AbsoluteDeterminantOneMatrices}
+    Orthogonal{T} = GeneralUnitaryMultiplicationGroup{T,ℝ,AbsoluteDeterminantOneMatrixType}
 
 Orthogonal group $\mathrm{O}(n)$ represented by [`OrthogonalMatrices`](@ref).
 
@@ -7,7 +7,8 @@ Orthogonal group $\mathrm{O}(n)$ represented by [`OrthogonalMatrices`](@ref).
 
     Orthogonal(n::Int; parameter::Symbol=:type)
 """
-const Orthogonal{T} = GeneralUnitaryMultiplicationGroup{T,ℝ,AbsoluteDeterminantOneMatrices}
+const Orthogonal{T} =
+    GeneralUnitaryMultiplicationGroup{T,ℝ,AbsoluteDeterminantOneMatrixType}
 
 function Orthogonal(n::Int; parameter::Symbol=:type)
     return GeneralUnitaryMultiplicationGroup(OrthogonalMatrices(n; parameter=parameter))